Diagrams.Backend.Cairo.Raster

Description

Draw raster/bitmap images for the Diagrams Cairo backend.

High-level API

cairoRaster ∷ (Int → Int → CairoColor) → Int → Int → IO (Diagram Cairo R2)

Create an image "from scratch" by specifying the pixel data.

cairoLoadImage ∷ FilePath → IO (Diagram Cairo R2)

Load an image for Cairo (JPG, PNG, BMP, TIFF ...). The image is loaded once and which can be useful for e.g. tiles.

Colors

type CairoColor = Word32

An ARGB color as understood by Cairo. Cairo colors are with pre-multiplied alpha in a native endian format with the bytes from low to high: blue, green, red and alpha.

crgb ∷ Integral a ⇒ a → a → a → CairoColor

Construct a CairoColor from red, green and blue between [0 .. 255].

crgba ∷ Integral a ⇒ a → a → a → a → CairoColor

Construct a CairoColor from non-premultiplied red, green, blue and alpha between [0 .. 255]. Non-premultiplied colors means that translucent red is crgba 255 0 0 127.

crgbap ∷ Integral a ⇒ a → a → a → a → CairoColor

Construct a CairoColor from already premultiplied red, green, blue and alpha between [0 .. min alpha 255]. Premultiplied colors means that translucent red is crgbap 127 0 0 127.